+++ be/src/category/category.service.ts
@@ -1,4 +1,4 @@
-import { ForbiddenException, Injectable } from "@nestjs/common";
+import { ForbiddenException, Injectable, InternalServerErrorException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { CategoryRepository } from "./category.repository";
 import { UserEntity } from "src/user/entity/user.entity";
@@ -6,6 +6,7 @@ import { AddCategoryDto } from "./dto/add-category.dto";
 import { CategoryEntity } from "./entity/category.entity";
 import { DeleteCategoryDto } from "./dto/delete-category.dto";
 import { UpdateCategoryDto } from "./dto/update-category.dto";
+import { ulid } from "ulid";
 
 @Injectable()
 export class CategoryService {
@@ -15,15 +16,39 @@ export class CategoryService {
   ) {}
 
   async addCategory(dto: AddCategoryDto, user: UserEntity): Promise<string> {
-    return await this.categoryRepository.add(dto, user);
+    const { categoryName } = dto;
+    const categoryUuid = ulid();
+    const category = this.categoryRepository.create({
+      categoryUuid: categoryUuid,
+      categoryName: categoryName,
+      user: user,
+    });
+
+    try {
+      await this.categoryRepository.save(category);
+      return categoryUuid;
+    } catch (error) {
+      throw new InternalServerErrorException();
+    }
   }
 
-  async getCategoryEntity(categoryUuid: string): Promise<CategoryEntity | null> {
-    return await this.categoryRepository.checkByCategoryUuid(categoryUuid);
+  async getCategoryEntity(categoryUuid: string): Promise<CategoryEntity> {
+    return await this.categoryRepository.findOneByCategoryUuid(categoryUuid);
   }
 
   async deleteCategory(dto: DeleteCategoryDto) {
-    return await this.categoryRepository.deleteCategory(dto);
+    const { categoryUuid } = dto;
+
+    try {
+      const record = await this.categoryRepository.findOne({ where: { categoryUuid }, relations: ["scheduleMeta"] });
+      record.scheduleMeta.forEach((entity) => {
+        entity.softRemove();
+      });
+
+      await this.categoryRepository.softDelete({ categoryUuid });
+    } catch (e) {
+      throw new InternalServerErrorException();
+    }
   }
 
   async updateCategory(dto: UpdateCategoryDto) {
